+++ server/executor/pollingprofile/polling_profile_resource.go
@@ -9,7 +9,7 @@ import (
 	"math"
 	"time"
 
-	"github.com/kubeshop/tracetest/server/id"
+	"github.com/kubeshop/tracetest/server/pkg/id"
 	"github.com/kubeshop/tracetest/server/resourcemanager"
 )
