+++ server/model/run.go
@@ -5,7 +5,7 @@ import (
 	"math"
 	"time"
 
-	"github.com/kubeshop/tracetest/server/id"
+	"github.com/kubeshop/tracetest/server/pkg/id"
 	"github.com/kubeshop/tracetest/server/pkg/maps"
 )
